# Faster checkReservedWord